+++ bidscoin/Modules/MRI/__init__.py
@@ -1,0 +1,4 @@
+from .DICOM import DICOM
+from .Nifti_SPM12 import Nifti_SPM12
+
+__all__ = ["DICOM", "Nifti_SPM12"]