+++ bidscoin/bidsmap/__init__.py
@@ -1,0 +1,2 @@
+from ._run import Run
+from ._bidsmap import Bidsmap